Moving Stage types to tests as they are only used there (and will not use it when fixed)
